# RFC: Improve some type definitions and decrease specialization