@timholy 2021-08-07
One could add a local version of `inferencebarrier` to support Julia 1.0. https://github.com/JuliaLang/julia/blob/4931faa34a8a1c98b39fb52ed4eb277729120128/base/essentials.jl#L758