@JonasIsensee 2021-08-09
Hi @timholy ,
thanks so much for looking into this.
Your suggestions look promising.
Right now, I don't have the time to pursue this in depth but I'll come back to it in a few weeks.
If anyone else wants to play with it in the mean-time, I'm available for discussion.