@thchr 2021-08-09
Many thanks from here as well @timholy - a bunch to learn from studying your techniques here..! I had been playing around briefly with a simpler "workload", namely just testing what took time in saving/loading some simple data structures:
```jl
using SnoopCompile, Profile, JLD2

function workload()
    jldsave("example.jld2";
                a=1, b=2.0,  c=[1.0,2.0], d=rand(ComplexF64,2000,2000),
                e=Dict(s => 2 for s in ["a", "b", "c"]))
    x = Vector{Any}(undef, 6)
    jldopen("example.jld2") do file
        x[1], x[2], x[3], x[4], x[5] = file["a"], file["b"], file["c"], file["d"], file["e"]
    end
    return x
end

tinf = @snoopi_deep @profile workload()
```

Followed by:
```jl
using PyPlot
mref, ax = pgdsgui(tinf; by=exclusive)
```

I revisited this now with your commit as comparison. Playing around with the slightly simpler workload highlighted many of the same methods as I can see you have touched here (`write_data`, `write_dataset`, `read_scalar`, `eltype`, `commit`, etc.) - but also highlighted `getproperty` and `setproperty` which surprised me a little. I'll need to study your tricks in the commit here to understand the actions you take to remedy the problems; but it's a very nice and complete casestudy - thanks again!

Thanks also bunch for taking the time to explain your further in #340 and in particular for pointing out that:
- I should remove the precompile statements before trying the despecialization; makes a bunch of sense...
- The `by=exclusive` kwarg to `pgdsgui`

One quick question (if you have the time): in #340, you had a line `Profile.init(delay=0.01, n=10^7)`, i.e. you increase the time delay between backtraces - what's the motivation for this? Should one generally be reducing this for SnoopCompile, or?

As a practical matter, I think this seems very worthwhile to do - and also to remove the precompile statements I had added (I saw your ["prioritized steps list"](https://github.com/JuliaSIMD/TriangularSolve.jl/pull/8#issuecomment-895398355); makes sense!).