@timholy 2021-08-09
> but also highlighted getproperty and setproperty which surprised me a little.

These you need to ignore. Every `struct` that has an `obj.x` access calls `getproperty` and [constant-specializes](https://compileroptimizations.com/category/constant_propagation.htm) them, but they show up as inference workloads. No good way around it short of making inference faster (which may be difficult/impossible), I'm afraid.

> you increase the time delay between backtraces - what's the motivation for this?

Any time you want to collect more snapshots than there is room for in the profile buffer, you might consider this. I typically leave it where it is until a workload takes ~10s, and then I start dropping it. When running the DataFrames test suite I only take a snapshot every 100ms.